Produce cell output buffers
